pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alpaca-trader"
version = "0.1.0"
description = "An automated trading bot built with Python and Alpaca SDK"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "alpaca-trade-api>=2.3.0",
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "python-dotenv>=0.19.0",
    "apscheduler>=3.8.0",
    "pytz>=2021.1",
    "loguru>=0.5.3",
    "matplotlib>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
]

# ML-related optional dependencies
ml = [
    "scikit-learn>=1.0.0",
    "tensorflow>=2.8.0",
    "keras>=2.8.0",
    "torch>=1.10.0",
    "xgboost>=1.5.0",
    "lightgbm>=3.3.0",
    "statsmodels>=0.13.0",
]

# NLP-related optional dependencies
nlp = [
    "nltk>=3.6.0",
    "transformers>=4.16.0",
    "spacy>=3.2.0",
]

# Reinforcement learning optional dependencies
rl = [
    "gym>=0.21.0",
    "stable-baselines3>=1.5.0",
]

# All ML-related dependencies
ml-all = [
    "scikit-learn>=1.0.0",
    "tensorflow>=2.8.0",
    "keras>=2.8.0",
    "torch>=1.10.0",
    "xgboost>=1.5.0",
    "lightgbm>=3.3.0",
    "statsmodels>=0.13.0",
    "nltk>=3.6.0",
    "transformers>=4.16.0",
    "spacy>=3.2.0",
    "gym>=0.21.0",
    "stable-baselines3>=1.5.0",
]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest]
testpaths = ["tests"]